refactor: UIAlertController 하나로 통일
